Commit #02:
(1) Likelihood Models
* Logistic Regression with different feature sets were implemented:
** Product views counters & Action (one hot encoding/vector);
** Product views frequencies & Action (one hot encoding/vector);
** Product views frequencies & Action (one hot encoding/vector) *with* Flag of the presence of an Action in Views;
** Product views frequencies & Action (one hot encoding/vector) *with* cross-fit (2nd polynomial);
** Product views frequencies & Action (one hot encoding/vector) *with* cross-fit (2nd polynomial) *with* Weight function based on time of an Event.
* Comparision of Agents based on Logistic Regression vs Organic Event Counter Agent was provided.

(2) Improvements
* Some adjustments in logging of data to confirm the notations used in the article of RecoGym.
* RandomState is shared among Agent and Environment to guarantee reproducibility.
* User ID is counted from `0' to make the code consistent with the rest of notions like Product, Time.

(3) New Features
* Omega can either be updated/not updated for Bandit Events (controlled by flag: 'change_omega_for_bandits').

+++ reco_gym/envs/reco_env_v1.py
@@ -95,9 +95,11 @@ class RecoEnv1(AbstractEnv):
         log_uprob = matmul(self.Gamma, self.omega)[:, 0] + self.mu_organic
         log_uprob = log_uprob - max(log_uprob)
         uprob = exp(log_uprob)
-        self.product_view = self.rng.choice(
-            self.num_products,
-            p = uprob / sum(uprob)
+        self.product_view = int(
+            self.rng.choice(
+                self.num_products,
+                p = uprob / sum(uprob)
+            )
         )
 
     def generate_beta(self, number_of_flips):
@@ -135,4 +137,3 @@ class RecoEnv1(AbstractEnv):
 
         self.beta = self.Gamma[index, :]
         self.mu_bandit = self.mu_organic[index]
-
